+++ lib/event_bus/notifier.ex
@@ -9,17 +9,24 @@ defmodule EventBus.Notifier do
   use GenServer
   alias EventBus.Model.Event
 
-  @backend Application.get_env(:event_bus, :notifier_backend,
-    EventBus.Service.Notifier)
+  @backend Application.get_env(
+             :event_bus,
+             :notifier_backend,
+             EventBus.Service.Notifier
+           )
 
   @doc false
   def start_link,
     do: GenServer.start_link(__MODULE__, nil, name: __MODULE__)
 
+  @doc false
+  def init(args),
+    do: {:ok, args}
+
   @doc """
   Notify event to its listeners
   """
-  @spec notify(Event.t) :: no_return()
+  @spec notify(Event.t()) :: no_return()
   def notify(%Event{} = event),
     do: GenServer.cast(__MODULE__, {:notify, event})
 
@@ -28,7 +35,7 @@ defmodule EventBus.Notifier do
   ###########################################################################
 
   @doc false
-  @spec handle_cast({:notify, Event.t}, nil) :: no_return()
+  @spec handle_cast({:notify, Event.t()}, nil) :: no_return()
   def handle_cast({:notify, event}, state) do
     @backend.notify(event)
     {:noreply, state}