+++ lib/event_bus/services/subscription.ex
@@ -12,6 +12,7 @@ defmodule EventBus.Service.Subscription do
   def subscribe({listener, topics}) do
     {listeners, topic_map} = load_state()
     listeners = add_or_update_listener(listeners, {listener, topics})
+
     topic_map =
       topic_map
       |> add_listener_to_topic_map({listener, topics})
@@ -25,6 +26,7 @@ defmodule EventBus.Service.Subscription do
   def unsubscribe(listener) do
     {listeners, topic_map} = load_state()
     listeners = List.keydelete(listeners, listener, 0)
+
     topic_map =
       topic_map
       |> remove_listener_from_topic_map(listener)
@@ -34,11 +36,12 @@ defmodule EventBus.Service.Subscription do
   end
 
   @doc false
-  @spec register_topic(String.t | atom()) :: no_return()
+  @spec register_topic(String.t() | atom()) :: no_return()
   def register_topic(topic) do
     {listeners, topic_map} = load_state()
+
     topic_subscribers =
-      Enum.reduce(listeners, [], fn({listener, topics}, acc) ->
+      Enum.reduce(listeners, [], fn {listener, topics}, acc ->
         if RegexUtil.superset?(topics, topic), do: [listener | acc], else: acc
       end)
 
@@ -46,7 +49,7 @@ defmodule EventBus.Service.Subscription do
   end
 
   @doc false
-  @spec unregister_topic(String.t | atom()) :: no_return()
+  @spec unregister_topic(String.t() | atom()) :: no_return()
   def unregister_topic(topic) do
     {listeners, topic_map} = load_state()
     save_state({listeners, Map.drop(topic_map, [topic])})
@@ -57,6 +60,7 @@ defmodule EventBus.Service.Subscription do
     {listeners, _topic_map} = load_state()
     listeners
   end
+
   def subscribers(topic) do
     {_listeners, topic_map} = load_state()
     topic_map[topic] || []
@@ -72,6 +76,7 @@ defmodule EventBus.Service.Subscription do
   defp add_listener_to_topic_map(topic_map, {listener, topics}) do
     Enum.map(topic_map, fn {topic, topic_listeners} ->
       topic_listeners = List.delete(topic_listeners, listener)
+
       if RegexUtil.superset?(topics, topic) do
         {topic, [listener | topic_listeners]}
       else
@@ -96,6 +101,7 @@ defmodule EventBus.Service.Subscription do
 
   defp init_topic_map do
     topics = Topic.all()
+
     topics
     |> Enum.map(fn topic -> {topic, []} end)
     |> Enum.into(%{})