+++ test/event_bus/services/subscription_test.exs
@@ -1,18 +1,24 @@
 defmodule EventBus.Service.SubscriptionTest do
   use ExUnit.Case, async: false
-  alias EventBus.Support.Helper.{InputLogger, Calculator, MemoryLeakerOne,
-    AnotherCalculator}
+
+  alias EventBus.Support.Helper.{
+    InputLogger,
+    Calculator,
+    MemoryLeakerOne,
+    AnotherCalculator
+  }
+
   alias EventBus.Service.Subscription
 
   doctest Subscription
 
   setup do
-    on_exit fn ->
+    on_exit(fn ->
       Subscription.unregister_topic(:auto_subscribed)
       Subscription.unregister_topic(:metrics_received)
       Subscription.unregister_topic(:metrics_summed)
       Process.sleep(100)
-    end
+    end)
 
     Subscription.register_topic(:auto_subscribed)
     Subscription.register_topic(:metrics_received)
@@ -35,11 +41,11 @@ defmodule EventBus.Service.SubscriptionTest do
     Process.sleep(300)
 
     assert [
-      {AnotherCalculator, [".*"]},
-      {{MemoryLeakerOne, %{}}, [".*"]},
-      {{Calculator, %{}}, [".*"]},
-      {{InputLogger, %{}}, [".*"]}
-    ] == Subscription.subscribers()
+             {AnotherCalculator, [".*"]},
+             {{MemoryLeakerOne, %{}}, [".*"]},
+             {{Calculator, %{}}, [".*"]},
+             {{InputLogger, %{}}, [".*"]}
+           ] == Subscription.subscribers()
   end
 
   test "does not subscribe same listener" do
@@ -61,7 +67,7 @@ defmodule EventBus.Service.SubscriptionTest do
     Process.sleep(300)
 
     assert [{{MemoryLeakerOne, %{}}, [".*"]}, {{InputLogger, %{}}, [".*"]}] ==
-      Subscription.subscribers()
+             Subscription.subscribers()
   end
 
   test "register_topic auto subscribe workers" do
@@ -77,7 +83,7 @@ defmodule EventBus.Service.SubscriptionTest do
     Process.sleep(300)
 
     assert [{InputLogger, %{}}, {Calculator, %{}}, AnotherCalculator] ==
-      Subscription.subscribers(topic)
+             Subscription.subscribers(topic)
   end
 
   test "unregister_topic delete subscribers" do
@@ -106,27 +112,26 @@ defmodule EventBus.Service.SubscriptionTest do
     Subscription.subscribe({{InputLogger, %{}}, [".*"]})
     Process.sleep(300)
 
-    assert [{InputLogger, %{}}] ==
-      Subscription.subscribers(:metrics_received)
-    assert [{InputLogger, %{}}] ==
-      Subscription.subscribers(:metrics_summed)
+    assert [{InputLogger, %{}}] == Subscription.subscribers(:metrics_received)
+    assert [{InputLogger, %{}}] == Subscription.subscribers(:metrics_summed)
   end
 
   test "subscribers with event type and without config" do
     Subscription.subscribe({AnotherCalculator, [".*"]})
     Process.sleep(300)
 
-    assert [AnotherCalculator] ==
-      Subscription.subscribers(:metrics_received)
-    assert [AnotherCalculator] ==
-      Subscription.subscribers(:metrics_summed)
+    assert [AnotherCalculator] == Subscription.subscribers(:metrics_received)
+    assert [AnotherCalculator] == Subscription.subscribers(:metrics_summed)
   end
 
   test "state persistency to Application environment" do
-    Subscription.subscribe({{InputLogger, %{}}, ["metrics_received",
-      "metrics_summed"]})
+    Subscription.subscribe(
+      {{InputLogger, %{}}, ["metrics_received", "metrics_summed"]}
+    )
+
     Subscription.subscribe({AnotherCalculator, ["metrics_received$"]})
     Process.sleep(300)
+
     expected = {
       [
         {AnotherCalculator, ["metrics_received$"]},